requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
line-bot-sdk==2.2.1
MarkupSafe==2.1.1
multidict==6.0.2
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.1
yarl==1.7.2
